requirements.txt
pandas
numpy
fpdf2
fonttools
requests
